# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TEST OpenApi  Документація",
        "description": "L5 Swagger OpenApi опис",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/goods": {
            "get": {
                "tags": [
                    "Goods"
                ],
                "summary": "Сторінка клієнтів",
                "description": "Повертає json список клієнтів",
                "operationId": "094b5e9ccb5dc4a980fea10f2d6136e6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/goods/store": {
            "post": {
                "tags": [
                    "Goods"
                ],
                "summary": "Створити нового клієнта",
                "description": "Повертає статус",
                "operationId": "1078d6c0b1966e4c247d6c8819394b21",
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Андрій"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Мельник"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+38(095)-022-81-26"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "vagary18"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/goods/show/{id}": {
            "get": {
                "tags": [
                    "Goods"
                ],
                "summary": "Дані одного клієнта",
                "description": "Повертає json",
                "operationId": "f5ef65ad6702f9b5988652dd1b5ea432",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/goods/update/{id}": {
            "put": {
                "tags": [
                    "Goods"
                ],
                "summary": "Зміна даних клієнта",
                "description": "Повертає статус",
                "operationId": "3e0c50481a2388c7f921aaf643265fe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id акаунта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "_bigmir_smile_"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+38(095)-022-81-26"
                                    },
                                    "email_password": {
                                        "type": "string",
                                        "example": "vagary18"
                                    },
                                    "session": {
                                        "type": "string",
                                        "example": "49943231002%3AyrQ4nK302XxHUW%3A18"
                                    },
                                    "bussines_instagram_account_id": {
                                        "type": "string",
                                        "example": "17841444782304448"
                                    },
                                    "date_created": {
                                        "type": "string",
                                        "example": "2022-05-18"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "vagary18"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ruslan.orelec123@gmail.com"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "proxy_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "user_agent_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/goods/destroy/{id}": {
            "delete": {
                "tags": [
                    "Goods"
                ],
                "summary": "Видалити клієнта",
                "description": "Повертає статус",
                "operationId": "54a50089d8f99dc451889fd408710961",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}